attempt to make cmake work

+++ modules/compute/ComputeLasLoader.h
@@ -68,7 +68,7 @@ struct ComputeLasData : public Resource {
 		}
 
 		//numPoints = min(numPoints, 50'000'000ll);
-		numPoints = min(numPoints, 1'000'000'000ll);
+		numPoints = min((long long int) numPoints, 1'000'000'000ll);
 		// 1'073'741'824
 
 		data->offsetToPointData = buffer_header->get<uint32_t>(96);
